# backend-go/internal/infrastructure/http/apidocs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Course AI API",
    "version": "1.0.0",
    "description": "API HTTP du MVP Course AI protégée par des sessions Clerk. La génération complète est asynchrone et s'appuie sur des jobs PostgreSQL durables."
  },
  "servers": [
    {
      "url": "/",
      "description": "Hôte courant"
    }
  ],
...
  "paths": {
...
    "/api/generations": {
      "get": {
        "tags": ["Generations"],
        "operationId": "listGenerationRequests",
        "summary": "Lister les demandes de génération",
        "description": "Retourne l'historique paginé du propriétaire authentifié, y compris les demandes en attente de clarification ou sans cours créé.",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {"$ref": "#/components/schemas/PipelineStatus"}
          },
          {
            "name": "page",
            "in": "query",
            "schema": {"type": "integer", "minimum": 1, "default": 1}
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {"type": "integer", "minimum": 1, "maximum": 100, "default": 20}
          }
        ],
        "responses": {
          "200": {
            "description": "Page de demandes de génération.",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GenerationPageResponse"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "429": {"$ref": "#/components/responses/RateLimited"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"},
          "500": {"$ref": "#/components/responses/InternalError"}
        }
      },
      "post": {
        "tags": [
          "Generations"
        ],
        "operationId": "startFullGeneration",
        "summary": "Démarrer une génération complète",
        "description": "Persiste la demande et enfile un job d'analyse. La réponse 202 est immédiate. Si l'analyse requiert des précisions, la pipeline passe à awaiting_clarification et le client doit appeler l'URL fournie dans actionRequired. Dans le cas contraire, le job d'architecture est automatiquement enfilé.",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Clé stable recommandée pour rejouer la création sans dupliquer la demande. Réutiliser une clé avec un autre prompt retourne 409.",
            "schema": {
              "type": "string",
              "maxLength": 255
            },
            "example": "course-linux-2026-08-18"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartGenerationRequest"
              },
              "example": {
                "prompt": "Je veux une formation sur Linux pour administrer des serveurs."
              }
            }
          }
        },
        "responses": {
          "202": {
            "$ref": "#/components/responses/GenerationAccepted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/generations/{requestID}": {
      "delete": {
        "tags": ["Generations"],
        "operationId": "deleteGenerationRequest",
        "summary": "Supprimer une demande et toutes ses données",
        "description": "Supprime atomiquement la demande possédée par l'utilisateur, ses jobs, son cours et tout le graphe pédagogique par cascade PostgreSQL.",
        "parameters": [{"$ref": "#/components/parameters/RequestID"}],
        "responses": {
          "204": {"description": "Demande supprimée."},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "500": {"$ref": "#/components/responses/InternalError"}
        }
      }
    },
...
    "/api/generations/{requestID}/status": {
      "get": {
        "tags": [
          "Generations"
        ],
        "operationId": "getGenerationStatus",
        "summary": "Consulter l'état fonctionnel d'une génération",
        "description": "Endpoint de polling principal. actionRequired indique au frontend qu'une soumission de clarifications est nécessaire.",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestID"
          }
        ],
        "responses": {
          "200": {
            "description": "État courant de la pipeline.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationStatusResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/generations/{requestID}/result": {
      "get": {
        "tags": [
          "Generations"
        ],
        "operationId": "getGenerationResult",
        "summary": "Récupérer le résultat complet d'une génération",
        "description": "Retourne la demande et le graphe complet course/modules/lessons/activities. Tant que la pipeline n'est pas completed, l'API retourne 409.",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestID"
          }
        ],
        "responses": {
          "200": {
            "description": "Formation complète générée.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationResultResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/generations/{requestID}/jobs": {
      "get": {
        "tags": ["Generation jobs"],
        "operationId": "listGenerationJobs",
        "summary": "Retrouver les jobs d'une generation",
        "description": "Retourne les etats publics des jobs de la demande appartenant a l'utilisateur connecte, pour reprendre le suivi apres rechargement. Les payloads internes ne sont pas exposes.",
        "parameters": [{"$ref": "#/components/parameters/RequestID"}],
        "responses": {
          "200": {"description": "Jobs de la demande.", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/GenerationJobResponse"}}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "500": {"$ref": "#/components/responses/InternalError"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
    "/api/generation-jobs/{jobID}": {
      "get": {
        "tags": [
          "Generation jobs"
        ],
        "operationId": "getGenerationJob",
        "summary": "Consulter l'état technique d'un job",
        "description": "Expose le type, les tentatives, le verrou et l'erreur terminale d'un job asynchrone.",
        "parameters": [
          {
            "$ref": "#/components/parameters/JobID"
          }
        ],
        "responses": {
          "200": {
            "description": "État du job.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationJobResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "RequestID": {
        "name": "requestID",
        "in": "path",
        "required": true,
        "description": "Identifiant UUID de la demande de génération.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "JobID": {
        "name": "jobID",
        "in": "path",
        "required": true,
        "description": "Identifiant UUID du job de génération.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "responses": {
      "GenerationAccepted": {
        "description": "Job durable accepté et enfilé.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GenerationStartedResponse"
            },
            "example": {
              "jobId": "953d5c51-337f-4e13-bf20-c5299f06ec14",
              "requestId": "d8c1571f-380e-44da-be06-9de1f8ba2aff",
              "status": "queued",
              "jobStatus": "queued",
              "statusUrl": "/api/generations/d8c1571f-380e-44da-be06-9de1f8ba2aff/status",
              "jobStatusUrl": "/api/generation-jobs/953d5c51-337f-4e13-bf20-c5299f06ec14",
              "resultUrl": "/api/generations/d8c1571f-380e-44da-be06-9de1f8ba2aff/result"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Corps, paramètre ou règle de validation invalide.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "code": "validation_error",
              "message": "invalid level"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Session Clerk absente ou invalide.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "code": "unauthenticated",
              "message": "authentication is required"
            }
          }
        }
      },
      "NotFound": {
        "description": "Ressource absente.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "code": "course_not_found",
              "message": "course not found"
            }
          }
        }
      },
      "Conflict": {
        "description": "L'état courant de la ressource interdit l'opération ou la clé d'idempotence est en conflit.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "code": "generation_not_completed",
              "message": "generation is not completed"
            }
          }
        }
      },
...
      "ServiceUnavailable": {
        "description": "Une dépendance applicative nécessaire est indisponible.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "code": "service_unavailable",
              "message": "generation service is unavailable"
            }
          }
        }
      },
      "PayloadTooLarge": {
        "description": "Le body JSON dépasse la limite configurée.",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"},
            "example": {"code": "payload_too_large", "message": "request body is too large"}
          }
        }
      },
      "RateLimited": {
        "description": "La limite de commandes de génération est atteinte.",
        "headers": {
          "Retry-After": {"schema": {"type": "integer"}, "description": "Délai conseillé en secondes."}
        },
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"},
            "example": {"code": "rate_limit_exceeded", "message": "too many generation commands"}
          }
        }
      },
      "InternalError": {
        "description": "Erreur interne masquée au client et journalisée avec sa stack trace.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "code": "internal_error",
              "message": "internal server error"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Code stable exploitable par le client."
          },
          "message": {
            "type": "string",
            "description": "Message public sans détail interne."
          }
        }
      },
      "StartGenerationRequest": {
        "type": "object",
        "required": [
          "prompt"
        ],
        "properties": {
          "prompt": {
            "type": "string",
            "minLength": 1,
            "maxLength": 4000
          }
        }
      },
...
      "GenerationStartedResponse": {
        "type": "object",
        "required": [
          "jobId",
          "requestId",
          "status",
          "jobStatus",
          "statusUrl",
          "jobStatusUrl",
          "resultUrl"
        ],
        "properties": {
          "jobId": {
            "type": "string",
            "format": "uuid"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/PipelineStatus"
          },
          "jobStatus": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "statusUrl": {
            "type": "string",
            "format": "uri-reference"
          },
          "jobStatusUrl": {
            "type": "string",
            "format": "uri-reference"
          },
          "resultUrl": {
            "type": "string",
            "format": "uri-reference"
          }
        }
      },
      "GenerationPageResponse": {
        "type": "object",
        "required": ["items", "page", "pageSize", "totalItems", "totalPages", "hasNext", "hasPrevious"],
        "properties": {
          "items": {"type": "array", "items": {"$ref": "#/components/schemas/GenerationSummaryResponse"}},
          "page": {"type": "integer", "minimum": 1},
          "pageSize": {"type": "integer", "minimum": 1, "maximum": 100},
          "totalItems": {"type": "integer", "minimum": 0},
          "totalPages": {"type": "integer", "minimum": 0},
          "hasNext": {"type": "boolean"},
          "hasPrevious": {"type": "boolean"}
        }
      },
      "GenerationSummaryResponse": {
        "type": "object",
        "required": ["requestId", "courseId", "initialUserPrompt", "title", "pipelineStatus", "courseStatus", "currentStep", "progressPercent", "isOutOfScope", "failureMessage", "createdAt", "updatedAt"],
        "properties": {
          "requestId": {"type": "string", "format": "uuid"},
          "courseId": {"type": "string", "format": "uuid", "nullable": true},
          "initialUserPrompt": {"type": "string"},
          "title": {"type": "string"},
          "pipelineStatus": {"$ref": "#/components/schemas/PipelineStatus"},
          "courseStatus": {"allOf": [{"$ref": "#/components/schemas/CourseStatus"}], "nullable": true},
          "currentStep": {"type": "string", "nullable": true},
          "progressPercent": {"type": "integer", "minimum": 0, "maximum": 100},
          "isOutOfScope": {"type": "boolean"},
          "failureMessage": {"type": "string", "nullable": true},
          "createdAt": {"type": "string", "format": "date-time"},
          "updatedAt": {"type": "string", "format": "date-time"}
        }
      },
      "GenerationJobResponse": {
        "type": "object",
        "required": [
          "id",
          "requestId",
          "parentJobId",
          "kind",
          "status",
          "targetId",
          "priority",
          "attemptCount",
          "maxAttempts",
          "availableAt",
          "lockedUntil",
          "startedAt",
          "completedAt",
          "lastErrorCode",
          "lastErrorMessage",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "parentJobId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "kind": {
            "$ref": "#/components/schemas/JobKind"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "targetId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "priority": {
            "type": "integer"
          },
          "attemptCount": {
            "type": "integer",
            "minimum": 0
          },
          "maxAttempts": {
            "type": "integer",
            "minimum": 1
          },
          "availableAt": {
            "type": "string",
            "format": "date-time"
          },
          "lockedUntil": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastErrorCode": {
            "type": "string",
            "nullable": true
          },
          "lastErrorMessage": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "GenerationStatusResponse": {
        "type": "object",
        "required": [
          "requestId",
          "courseId",
          "pipelineStatus",
          "courseStatus",
          "currentStep",
          "progressPercent",
          "failureMessage",
          "isOutOfScope",
          "errorMessage",
          "warningMessage",
          "suggestedTitle",
          "shortSynopsis",
          "detectedCurrentLevel",
          "detectedTargetLevel",
          "detectedGoal",
          "detectedLanguage",
          "clarificationQuestions",
          "actionRequired"
        ],
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "courseId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "pipelineStatus": {
            "$ref": "#/components/schemas/PipelineStatus"
          },
          "courseStatus": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CourseStatus"
              }
            ],
            "nullable": true
          },
          "currentStep": {
            "type": "string",
            "nullable": true
          },
          "progressPercent": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "failureMessage": {
            "type": "string",
            "nullable": true
          },
          "isOutOfScope": {
            "type": "boolean"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "warningMessage": {
            "type": "string",
            "nullable": true
          },
          "suggestedTitle": {
            "type": "string",
            "nullable": true
          },
          "shortSynopsis": {
            "type": "string",
            "nullable": true
          },
          "detectedCurrentLevel": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Level"
              }
            ],
            "nullable": true
          },
          "detectedTargetLevel": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Level"
              }
            ],
            "nullable": true
          },
          "detectedGoal": {
            "type": "string",
            "nullable": true
          },
          "detectedLanguage": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Language"
              }
            ],
            "nullable": true
          },
          "clarificationQuestions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClarificationQuestionResponse"
            }
          },
          "actionRequired": {
            "allOf": [
              {
                "$ref": "#/components/schemas/GenerationActionRequiredResponse"
              }
            ],
            "nullable": true
          }
        }
      },
      "GenerationActionRequiredResponse": {
        "type": "object",
        "required": [
          "type",
          "url"
        ],
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "submit_clarifications"
            ]
          },
          "url": {
            "type": "string",
            "format": "uri-reference"
          }
        }
      },
      "GenerationResultResponse": {
        "type": "object",
        "required": [
          "request",
          "course"
        ],
        "properties": {
          "request": {
            "$ref": "#/components/schemas/GenerationRequestResponse"
          },
          "course": {
            "$ref": "#/components/schemas/CourseResponse"
          }
        }
      },
...
      "ClarificationQuestionResponse": {
        "type": "object",
        "required": [
          "id",
          "question",
          "options",
          "allowMultiple"
        ],
        "properties": {
          "id": {
            "type": "string",
            "enum": [
              "goals",
              "currentLevel",
              "targetLevel"
            ]
          },
          "question": {
            "type": "string"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClarificationOptionResponse"
            }
          },
          "allowMultiple": {
            "type": "boolean"
          }
        }
      },
      "ClarificationOptionResponse": {
        "type": "object",
        "required": [
          "value",
          "label"
        ],
        "properties": {
          "value": {
            "type": "string"
          },
          "label": {
            "type": "string"
          }
        }
      },
...
      "Language": {
        "type": "string",
        "enum": [
          "fr",
          "en"
        ]
      },
      "Level": {
        "type": "string",
        "enum": [
          "beginner",
          "intermediate",
          "advanced",
          "expert",
          "unknown"
        ]
      },
...
      "PipelineStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "awaiting_clarification",
          "completed",
          "failed"
        ]
      },
      "CourseStatus": {
        "type": "string",
        "enum": [
          "analysis_pending",
          "analysis_completed",
          "architecture_generating",
          "structure_generated",
          "lessons_generating",
          "lessons_generated",
          "content_generating",
          "completed",
          "failed"
        ]
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "retry_scheduled",
          "completed",
          "failed",
          "cancelled"
        ]
      },
      "JobKind": {
        "type": "string",
        "enum": [
          "analysis",
          "architecture",
          "lesson_plan",
          "lesson_content",
          "module_content",
          "finalize_course"
        ]
      }
    },
    "securitySchemes": {
      "ClerkBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Clerk session JWT",
        "description": "Session token Clerk transmis dans Authorization: Bearer <token>."
      }
    }
  },
  "security": [
    {
      "ClerkBearer": []
    }
  ]
}